command_center/requirements.txt
PyQt6>=6.5
requests>=2.28
python-dotenv>=1.0
google-genai>=1.0
anthropic>=0.30
